Refactored code
Fixed the way to store and retrieve objects correctly:
I was calling sparse data and trying to reach the ID of objects in an array without storing those object with ID at first.
To display the detailed information for each movie, the code makes separate API calls for each movie using its imdbID. The detailed movie data for each movie is returned from these separate API calls. It doesn't display movie data directly because of the asynchronous nature of the API. more in the comments of the iteration of researchArr in index.js